% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-setup-zip.R
\name{get_setup_zip}
\alias{get_setup_zip}
\title{Create Zipped Project}
\usage{
get_setup_zip(
  outfile,
  project_name = "new_project",
  dependencies_file_path = NULL,
  dependencies = NULL,
  add_gitzip = TRUE
)
}
\arguments{
\item{outfile}{file path for desired zipped project directory.}

\item{project_name}{name of new project. Defaults to "new_project".}

\item{dependencies_file_path}{if building project library based on a script,
specify the path of this file. It should be similar to a dependencies.R
script that you may find in a rhino app.}

\item{dependencies}{vector or list of packages.}

\item{add_gitzip}{whether to add a gitzip.sh file to new project directory.}
}
\value{
file path of new zipped project.
}
\description{
Create Zipped Project
}
\examples{
\donttest{
  new_proj_path <- "./test_proj.zip"
  new_proj_zip <- get_setup_zip(outfile = new_proj_path, dependencies = "glue")
}
}
